AdvancedFilterComponents and AdvancedMapComponents added. GRACE now uses these components.
