+++ src/app/cesium-map/advanced/grace/grace-legend.component.scss
