+++ src/app/cesium-map/csmap.component.scss
@@ -60,7 +60,7 @@
     color: white;
 }
 
-.toolbar {
+.advancedmapcomponent {
     position: absolute;
     z-index: 1;
 }
