+++ src/app/menupanel/common/filterpanel/advance/grace/grace-advanced-filter.component.html
@@ -1,4 +1,4 @@
-<div class="grace-toolbar">
+<div class="grace-advancedfilter">
     <div class="row">
         <!-- Time series graphs -->
         <div ngbDropdown>
