+++ src/app/menupanel/common/filterpanel/advance/grace/grace-advanced-filter.component.scss
@@ -1,5 +1,5 @@
 :host {
-    .grace-toolbar {
+    .grace-advancedfilter {
         margin-left: 33px;
         margin-right: 33px;
     }
