+++ src/app/menupanel/common/filterpanel/dynamic.advancefilter.component.ts
@@ -30,11 +30,11 @@ export class DynamicAdvancefilterComponent {
 
   loadComponent() {
 
-    if (!ref.advanceFilter[this._layer.id]) {
+    if (!ref.advancedFilter[this._layer.id]) {
       return;
     }
 
-    const componentFactory = this.componentFactoryResolver.resolveComponentFactory(ref.advanceFilter[this._layer.id]);
+    const componentFactory = this.componentFactoryResolver.resolveComponentFactory(ref.advancedFilter[this._layer.id]);
 
     const viewContainerRef = this.dyanmicAnalyticHost
     viewContainerRef.clear();
