+++ src/app/menupanel/data-explorer-record/data-explorer-record.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, ViewContainerRef, OnInit } from "@angular/core";
 import {  CSWRecordModel,  CsMapService,  ManageStateService,  UtilitiesService,  LayerModel} from "@auscope/portal-core-ui";
 import { NgbModal } from "@ng-bootstrap/ng-bootstrap";
-import { ToolbarComponentsService } from "app/services/ui/toolbar-components.service";
+import { AdvancedMapComponentService } from "app/services/ui/advanced-map-component.service";
 import { environment } from "environments/environment";
 import * as _ from "lodash";
 import { RecordModalComponent } from "../record-modal/record-modal.component";
@@ -28,7 +28,7 @@ export class DataExplorerRecordComponent implements OnInit {
   constructor(
     public csMapService: CsMapService,
     private manageStateService: ManageStateService,
-    private toolbarService: ToolbarComponentsService,
+    private advancedMapComponentService: AdvancedMapComponentService,
     public modalService: NgbModal
   ) {
     this.optionalFilters = [];
@@ -61,7 +61,7 @@ export class DataExplorerRecordComponent implements OnInit {
     if (cswRecord) {
       const modelRef = this.modalService.open(RecordModalComponent, {
         size: "lg",
-        backdrop:false
+        backdrop: false
       });
       modelRef.componentInstance.cswRecords = this.layer.cswRecords;
       modelRef.componentInstance.layer = this.layer;
@@ -143,8 +143,8 @@ export class DataExplorerRecordComponent implements OnInit {
       $("#sidebar-toggle-btn").click();
     }
 
-    // Add any toolbar components to map defined in refs.ts
-    this.toolbarService.addMapToolbarComponents(this.layer);
+    // Add any advanced map components to map defined in refs.ts
+    this.advancedMapComponentService.addAdvancedMapComponents(this.layer);
   }
 
   /**
@@ -156,8 +156,8 @@ export class DataExplorerRecordComponent implements OnInit {
     const layerModelList = this.csMapService.getLayerModelList();
     if (layerModelList.hasOwnProperty(layerId)) {
       this.csMapService.removeLayer(layerModelList[layerId]);
-      // Remove any layer specific toolbars
-      this.toolbarService.removeMapToolbarComponents(layerId);
+      // Remove any layer specific components
+      this.advancedMapComponentService.removeAdvancedMapComponents(layerId);
     }
   }
 }
