+++ src/app/modalwindow/querier/customanalytic/grace/grace-graph.modal.component.scss
