+++ src/app/services/wcustom/grace/grace.service.ts
@@ -2,8 +2,8 @@ import { environment } from '../../../../environments/environment';
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { BehaviorSubject, Observable } from 'rxjs';
-import { GraceStyleSettings } from '../../../toolbar/components/grace/grace-graph.models';
-import { first } from 'rxjs/operators';
+import { GraceStyleSettings } from '../../../modalwindow/querier/customanalytic/grace/grace-graph.models';
+import { serialize } from '@thi.ng/hiccup';
 
 /**
  * Service for the GRACE API
@@ -11,7 +11,7 @@ import { first } from 'rxjs/operators';
 @Injectable()
 export class GraceService {
     // Current GRACE style settings
-    private _currentGraceStyleSettings: BehaviorSubject<GraceStyleSettings> = new BehaviorSubject({
+    public currentGraceStyleSettings: GraceStyleSettings = {
         minColor: '#ff0000',
         minValue: -1,
         neutralColor: '#ffffff',
@@ -19,10 +19,10 @@ export class GraceService {
         maxColor: '#0000ff',
         maxValue: 1,
         transparentNeutralColor: false
-    });
-    public readonly currentGraceStyleSettings: Observable<GraceStyleSettings> = this._currentGraceStyleSettings.asObservable();
+    };
+    public currentGraceStyleSettingsBS: BehaviorSubject<GraceStyleSettings> = new BehaviorSubject(this.currentGraceStyleSettings);
     // Editor GRACE style settings
-    private _editedGraceStyleSettings: BehaviorSubject<GraceStyleSettings> = new BehaviorSubject({
+    public editedGraceStyleSettings: GraceStyleSettings = {
         minColor: '#ff0000',
         minValue: -1,
         neutralColor: '#ffffff',
@@ -30,8 +30,8 @@ export class GraceService {
         maxColor: '#0000ff',
         maxValue: 1,
         transparentNeutralColor: false
-    });
-    public readonly editedGraceStyleSettings: Observable<GraceStyleSettings> = this._editedGraceStyleSettings.asObservable();
+    };
+
     // Current GRACE epoch
     private _graceDate: BehaviorSubject<any> = new BehaviorSubject({ undefined });
     public readonly graceDate: Observable<Date> = this._graceDate.asObservable();
@@ -103,20 +103,93 @@ export class GraceService {
     }
 
     public setCurrentGraceStyleSettings(graceStyleSettings: GraceStyleSettings) {
-        this._currentGraceStyleSettings.next(graceStyleSettings);
+        this.currentGraceStyleSettings = graceStyleSettings;
+        this.currentGraceStyleSettingsBS.next(graceStyleSettings);
     }
 
     public setEditedGraceStyleSettings(graceStyleSettings: GraceStyleSettings) {
-        this._editedGraceStyleSettings.next(graceStyleSettings);
+        this.editedGraceStyleSettings = graceStyleSettings;
     }
 
     /**
-     * Make the editor style settings the current map settings
+     * Fetches the SLD_BODY parameter used to style a WMS request
+     *
+     * @method getGraceSld
+     * @return style sheet in string form
      */
-    public updateCurrentGraceStyleSettings() {
-        this.editedGraceStyleSettings.pipe(first()).subscribe(style => {
-            this._currentGraceStyleSettings.next(style);
-        });
+    public getGraceSld(): string {
+       this.setCurrentGraceStyleSettings(this.editedGraceStyleSettings);
+        const xmlHeader = serialize(['?xml', { 'version': '1.0', 'encoding': 'UTF-8' }]);
+        const styledLayerAttrs = {
+            'version': '1.0.0',
+            'xmlns:sld': 'http://www.opengis.net/sld',
+            'xmlns:ogc': 'http://www.opengis.net/ogc',
+            'xmlns:xsi': 'http://www.w3.org/2001/XMLSchema-instance',
+            'xsi:schemaLocation': 'http://www.opengis.net/sld http://schemas.opengis.net/sld/1.0.0/StyledLayerDescriptor.xsd'
+        };
+        const styledLayerDesc = (body: any) => ['sld:StyledLayerDescriptor', styledLayerAttrs, body]
+        const namedLayer = (body: string) => ['sld:NamedLayer', null, body];
+        const name = (nameStr: string) => ['sld:Name', null, nameStr];
+        const userStyle = (body: string) => ['sld:UserStyle', null, body];
+        const body1 = serialize(name('grace_style')) + this.getFeatureTypeStyle(this.currentGraceStyleSettings);
+        const body2 = serialize(name('grace:grace_view')) + serialize(userStyle(body1));
+        return xmlHeader + serialize(styledLayerDesc(namedLayer(body2)));
+    }
+
+    /**
+     * Assembles 'sld:FeatureTypeStyle' component of SLD_BODY parameter
+     *
+     * @method getFeatureTypeStyle
+     * @param styleSettings - the GRACE style settings
+     * @return XML 'sld:FeatureTypeStyle' string
+     */
+    private getFeatureTypeStyle(styleSettings: GraceStyleSettings): string {
+        const polygonSymbolizer = this.getPolySymbolizer(styleSettings);
+        const rule = ['sld:Rule', null, polygonSymbolizer];
+        return serialize(['sld:FeatureTypeStyle', null, rule]);
+    }
+
+    /**
+     * Assembles 'sld:PolygonSymbolizer' component of SLD_BODY parameter
+     *
+     * @method getPolySymbolizer
+     * @param styleSettings - the GRACE style settings
+     * @return XML 'sld:PolygonSymbolizer' string
+     */
+    private getPolySymbolizer(styleSettings: GraceStyleSettings): string {
+        const propertyName = ['ogc:PropertyName', null, 'ewh'];
+        const literalMinVal = ['ogc:Literal', null, styleSettings.minValue];
+        const literalMinCol = ['ogc:Literal', null, styleSettings.minColor];
+        const literalNeutralVal = ['ogc:Literal', null, styleSettings.neutralValue];
+        const literalNeutralCol = ['ogc:Literal', null, styleSettings.neutralColor];
+        const literalMaxVal = ['ogc:Literal', null, styleSettings.maxValue];
+        const literalMaxCol = ['ogc:Literal', null, styleSettings.maxColor];
+        const literalColorType = ['ogc:Literal', null, 'color'];
+        const func = ['ogc:Function', { 'name': 'Interpolate' },
+            [propertyName, literalMinVal, literalMinCol, literalNeutralVal,
+                literalNeutralCol, literalMaxVal, literalMaxCol, literalColorType]];
+        const fillCss = ['sld:CssParameter', { 'name': 'fill' }, func];
+
+        let fillCssOpacity = [];
+        if (styleSettings.transparentNeutralColor) {
+            const literalOpaqueVal = ['ogc:Literal', null, 1];
+            const literalTransparentVal = ['ogc:Literal', null, 0];
+            const literalNumericType = ['ogc:Literal', null, 'numeric'];
+            const opacityFillFunc = ['ogc:Function', { 'name': 'Interpolate' },
+            [propertyName, literalMinVal, literalOpaqueVal, literalNeutralVal,
+                literalTransparentVal, literalMaxVal, literalOpaqueVal, literalNumericType]];
+            fillCssOpacity = ['sld:CssParameter', {'name': 'fill-opacity'}, opacityFillFunc];
+        }
+        const fill = ['sld:Fill', null, [fillCss, fillCssOpacity]];
+
+        let stroke = [];
+        if (styleSettings.transparentNeutralColor === false) {
+            const strokeCss = ['sld:CssParameter', { 'name': 'stroke' }, func];
+            const strokeWidthCss = ['sld:CssParameter', { 'name': 'stroke-width' }, 1];
+            stroke = ['sld:Stroke', null, [strokeCss, strokeWidthCss]];
+        }
+
+        return serialize(['sld:PolygonSymbolizer', null, [fill, stroke]]);
     }
 
 }
