--- src/app/toolbar/toolbar.directive.ts
@@ -1,8 +1,0 @@
-import { Directive, ViewContainerRef } from '@angular/core';
-
-@Directive({
-  selector: '[toolbar]',
-})
-export class ToolbarDirective {
-  constructor(public viewContainerRef: ViewContainerRef) { }
-}
